feat: move retrospective code to module
